Primarily added some base functionality to the 'Order' tab.
Currently the user can add items to their basket, which updates the list of items the user has in their basket. The user can go back at any time and add another item(s) to their order, and the table will update. The basket total is also dynamic and reflects the total amount of the user's basket appropiately.

The Order tab however does not display the correct quantity for each item, and instead is returning 'nil', despite the quantity values being stored correctly, and can be printed - perhaps the code cannot be carried out correctly in the TableView delegate method?

Also, the 'Add' and 'Decrease' quantity buttons do not have any functionality yet.

- Refactored the rest of the custom cell classes into their own file.
- Removed some boiler-plate Google Firebase code.